requirements.txt
fastapi[standard]
asyncpg
websocket